+++ .github/workflows/build.yaml
@@ -1,0 +1,32 @@
+name: build
+
+on:
+  workflow_dispatch:
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  docker:
+    name: build
+    runs-on: ubuntu-20.04
+    timeout-minutes: 30
+    strategy:
+      fail-fast: false
+    steps:
+      - name: Check out code into the Go module directory
+        uses: actions/checkout@v2
+
+      - name: Setup Go
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.16.x
+
+      - name: Make help
+        run: sudo make help
+
+      - name: Build all cli
+        run: sudo make build
+
+      - name: Make images
+        run: sudo make image
