+++ .github/workflows/edgeadm.yaml
@@ -1,10 +1,10 @@
 name: edgeadm
 
-on: push
-#  workflow_dispatch:
-#  pull_request:
-#    branches:
-#      - main
+on:
+  workflow_dispatch:
+  pull_request:
+    branches:
+      - main
 
 jobs:
   docker:
@@ -25,7 +25,7 @@ jobs:
       - name: Setup Go
         uses: actions/setup-go@v2
         with:
-          go-version: 1.17.x
+          go-version: 1.16.x
 
       - name: Install edgeadm
         run: sudo make build BINS="lite-apiserver edgeadm" INSTALL_DIR=/usr/local/bin && yes | cp ./_output/linux/amd64/edgeadm /usr/local/bin/
