+++ .github/workflows/ut.yaml
@@ -1,14 +1,14 @@
 name: ut
 
-on: push
-#  workflow_dispatch:
-#  pull_request:
-#    branches:
-#      - main
+on:
+  workflow_dispatch:
+  pull_request:
+    branches:
+      - main
 
 jobs:
   docker:
-    name: e2e
+    name: ut
     runs-on: ubuntu-20.04
     timeout-minutes: 30
     strategy:
@@ -25,7 +25,7 @@ jobs:
       - name: Setup Go
         uses: actions/setup-go@v2
         with:
-          go-version: 1.17.x
+          go-version: 1.16.x
 
       - name: Install kubectl
         run: |
@@ -48,8 +48,11 @@ jobs:
           kubectl wait --for=condition=ready pods --namespace=kube-system -l k8s-app=kube-dns
           kubectl get nodes -o wide
 
-      - name: Run e2e
-        run: mkdir -p ~/gopath/src/github.com/superedge/ && PRJ_PREFIX="github.com/superedge/superedge" && go test ${PRJ_PREFIX}/pkg/ -v
+      - name: Run go mod
+        run: go mod tidy
+
+      - name: Run unit test
+        run: go test ./pkg/... -v
 
       - name: Delete cluster
         run: /usr/local/bin/kind delete cluster
