+++ pkg/edge-health/checkplugin/checkplugin.go
@@ -19,9 +19,9 @@ package checkplugin
 import (
 	"fmt"
 	"io/ioutil"
+	"k8s.io/klog/v2"
 	"net/http"
 	"sync"
-        "k8s.io/klog/v2"
 )
 
 type CheckPlugin interface {
