+++ pkg/edge-health/checkplugin/kubeletauthcheck.go
@@ -21,8 +21,8 @@ import (
 	"fmt"
 	"github.com/superedge/superedge/pkg/edge-health/common"
 	"github.com/superedge/superedge/pkg/edge-health/data"
+        "k8s.io/klog/v2"
 	"io/ioutil"
-	"k8s.io/klog/v2"
 	"net/http"
 	"strconv"
 	"strings"
@@ -36,6 +36,10 @@ type KubeletAuthCheckPlugin struct {
 
 func (p KubeletAuthCheckPlugin) Name() string {
 	return "KubeletAuthCheck"
+}
+
+func (p KubeletAuthCheckPlugin) Enable() bool {
+	return p.Enabled
 }
 
 func (p *KubeletAuthCheckPlugin) Set(s string) error {
@@ -66,9 +70,11 @@ func (p *KubeletAuthCheckPlugin) Set(s string) error {
 		}
 		(*p).PluginName = p.Name()
 	}
-	PluginInfo = NewPluginInfo()
-	PluginInfo.AddPlugin(p)
-	klog.V(4).Infof("len of plugins is %d", len(PluginInfo.Plugins))
+	if p.Enable() {
+		PluginInfo = NewPluginInfo()
+		PluginInfo.AddPlugin(p)
+		klog.V(4).Infof("len of plugins is %d", len(PluginInfo.Plugins))
+	}
 	return err
 }
