+++ pkg/statefulset-grid-daemon/hosts/hosts
@@ -1,3 +1,3 @@
 1.2.3.4 statefulsetgrid-demo-0.servicegrid-demo-svc.default.svc.cluster.local
-1.2.3.5 statefulsetgrid-demo-1.servicegrid-demo-svc.default.svc.cluster.local
-1.2.3.6 statefulsetgrid-demo-2.servicegrid-demo-svc.default.svc.cluster.local
+1.2.3.7 statefulsetgrid-demo-1.servicegrid-demo-svc.default.svc.cluster.local
+1.2.3.7 statefulsetgrid-demo-12.servicegrid-demo-svc.default.svc.cluster.local
