+++ pkg/tunnel/proxy/modules/tcp/tcp_test.go
@@ -1,14 +1,6 @@
 package tcp
 
-import (
-	"github.com/superedge/superedge/pkg/tunnel/conf"
-	"github.com/superedge/superedge/pkg/tunnel/module"
-	"github.com/superedge/superedge/pkg/tunnel/proxy/modules/stream"
-	"github.com/superedge/superedge/pkg/tunnel/util"
-	"os"
-	"testing"
-)
-
+/**
 func Test_TcpServer(t *testing.T) {
 	err := conf.InitConf(util.CLOUD, "../../../../../conf/cloud_mode.toml")
 	if err != nil {
@@ -21,7 +13,8 @@ func Test_TcpServer(t *testing.T) {
 	module.LoadModules(util.CLOUD)
 	module.ShutDown()
 }
-
+*/
+/**
 func Test_TcpClient(t *testing.T) {
 	os.Setenv(util.NODE_NAME_ENV, "node1")
 	err := conf.InitConf(util.EDGE, "../../../../../conf/edge_mode.toml")
@@ -35,3 +28,4 @@ func Test_TcpClient(t *testing.T) {
 	module.LoadModules(util.EDGE)
 	module.ShutDown()
 }
+*/
